+++ .vscode/settings.json
@@ -1,0 +1,19 @@
+{
+    "files.exclude": {
+        "**/.git": true,
+        "**/.svn": true,
+        "**/.hg": true,
+        "**/CVS": true,
+        "**/.DS_Store": true,
+        "**/Thumbs.db": true,
+        "package-lock.json": true,
+        ".vscode": true,
+        ".github": true,
+        "node_modules": true,
+        ".editorconfig": true,
+        ".gitignore": true,
+        ".jshintrc": true,
+        "LICENSE": true
+    },
+    "explorerExclude.backup": null
+}
