+++ input_templates/_email.js
@@ -1,12 +1,7 @@
+const validator_messages = require('./v_lidator_messages');
+
 const format = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
 
-const sysMsg = {
-  success: "✅ Success\n📫 Email verification successful." ,
-  error:  "❌Error\n📑 Email verification failed." ,
-};
-
 module.exports = (email) => {
-  const result = format.test(email);
-  if (process.consoleOutput === true) console.log((result === true) ? sysMsg.success : sysMsg.error);
-  return (result === true) ? true : sysMsg.error;
+  return (format.test(email)) ? validator_messages.password.success : validator_messages.password.error;
 };
