+++ input_templates/_username.js
@@ -1,27 +1,27 @@
+const validator_messages = require('./v_lidator_messages');
+
+var errorList = [];
 
 const _cfg = {
   min: 6,
   max: 32,
-  format: /^[A-Za-z][A-Za-z0-9_.]{0,255}$/
+  format: /^[A-Za-z][A-Za-z0-9_.]{0,255}$/,
+  validate_length = async (len) => {
+    if (len < _cfg.min) errorList.push(validator_messages.username.errors.min);
+    if (len > _cfg.max) errorList.push(validator_messages.username.errors.max);
+  },
+  validate_format = async (len) => {
+    if (!_cfg.format.test(username)) errorList.push(validator_messages.username.errors.chars);
+  },
 };
 
-const sysMsg = {
-  success: "✅ Success\n🤪 Username verification successful.",
-  errors: {
-    chars: "🙋‍♂️ Username can only have letters, numbers, underscore and dot.",
-    min: "🤯 Username minimum length is [ " + _cfg.min + " ]",
-    max: "💥 Username maximum length is [ " + _cfg.max + " ]",
-  }
-};
 
-module.exports = (username) => {
-  var errorList = [];
 
-  if (username.length < _cfg.min) errorList.push(sysMsg.errors.min);
-  if (username.length > _cfg.max) errorList.push(sysMsg.errors.max);
-  if (!_cfg.format.test(username)) errorList.push(sysMsg.errors.chars);
+module.exports = async (username) => {
 
-  if (process.consoleOutput === true) console.log(errorList.length === 0 ? sysMsg.success : JSON.stringify(errorList));
+  await _cfg.validate_length(String(username).length);
+  await _cfg.validate_format(username);
 
-  return (errorList.length === 0) ? true : { type: "ERROR", items: errorList };
+
+  return (errorList.length === 0) ? validator_messages.username.success : { type: "ERROR", items: errorList };
 };
