+++ input_templates/v_lidator.js
@@ -1,0 +1,7 @@
+const v_lidator = {
+    email: require('../_email'),
+    username: require('../_username'),
+    password: require('../_password')
+};
+
+module.exports = v_lidator;
